+++ src/Extended.Dapper.Core/Database/Entities/BaseEntity.cs
@@ -8,7 +8,7 @@ namespace Extended.Dapper.Core.Database.Entities
     {
         [Key]
         [AutoValue]
-        public int Id { get; set; }
+        public virtual int Id { get; set; }
 
         public bool Equals(BaseEntity other)
         {