+++ src/Extended.Dapper.Core/Sql/Generator/SqlGenerator.Insert.cs
@@ -1,10 +1,12 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using Extended.Dapper.Core.Attributes.Entities;
 using Extended.Dapper.Core.Attributes.Entities.Relations;
 using Extended.Dapper.Core.Database;
 using Extended.Dapper.Core.Mappers;
+using Extended.Dapper.Core.Sql.Metadata;
 using Extended.Dapper.Core.Sql.Query;
 using Extended.Dapper.Core.Sql.Query.Models;
 
@@ -40,32 +42,18 @@ namespace Extended.Dapper.Core.Sql.Generator
             insertQuery.Table = entityMap.TableName;
 
             // Grab all properties, except autovalue ones
-            var autoValueProperties = entityMap.PrimaryKeyPropertiesMetadata.Where(x => x.AutoValue);
+            var autoValueKeyProperties = entityMap.PrimaryKeyPropertiesMetadata.Where(x => x.AutoValue);
+            var autoValueProperties    = entityMap.AutoValuePropertiesMetadata;
 
-            foreach (var autoValueProperty in autoValueProperties)
-            {
-                var autoValueType = autoValueProperty.PropertyInfo.PropertyType;
-                var key = autoValueProperty.PropertyInfo.GetValue(entity);
-
-                if (!EntityMapper.IsKeyEmpty(key))
-                    insertQuery.IdAlreadyPresent = true;
-                else if (autoValueType == typeof(Guid))
-                    autoValueProperty.PropertyInfo.SetValue(entity, Guid.NewGuid());
-                else if (autoValueType == typeof(int))
-                {
-                    insertQuery.AutoIncrementKey = true;
-                    insertQuery.AutoIncrementField = autoValueProperty;
-                }
-                else
-                    throw new NotImplementedException($"AutoValue for type {autoValueProperty.PropertyInfo.PropertyType} is not supported");
-            }
+            this.FillAutoValueProperties(entity, insertQuery, autoValueProperties, false);
+            this.FillAutoValueProperties(entity, insertQuery, autoValueKeyProperties);
 
             insertQuery.Insert
                 .AddRange(entityMap.MappedPropertiesMetadata
                     .Where(x => !x.PropertyInfo.GetCustomAttributes<RelationAttributeBase>().Any()
                         && x.PropertyInfo.GetCustomAttribute<IgnoreOnInsertAttribute>() == null
                         && ( insertQuery.AutoIncrementField == null 
-                            || (insertQuery.AutoIncrementField != null && x.ColumnName != insertQuery.AutoIncrementField.ColumnName)))
+                            || (insertQuery.AutoIncrementField != null && x.PropertyName != insertQuery.AutoIncrementField.PropertyName)))
                     .Select(p => {
                         insertQuery.Params.Add("p_" + p.ColumnName, p.PropertyInfo.GetValue(entity));
 
@@ -74,6 +62,27 @@ namespace Extended.Dapper.Core.Sql.Generator
             ));
 
             return insertQuery;
+        }
+
+        protected void FillAutoValueProperties<T>(T entity, InsertSqlQuery insertQuery, IEnumerable<SqlPropertyMetadata> autoValueProperties, bool keyInsert = true)
+        {
+            foreach (var autoValueProperty in autoValueProperties)
+            {
+                var autoValueType = autoValueProperty.PropertyInfo.PropertyType;
+                var key = autoValueProperty.PropertyInfo.GetValue(entity);
+
+                if (keyInsert && !EntityMapper.IsKeyEmpty(key))
+                    insertQuery.IdAlreadyPresent = true;
+                else if (autoValueType == typeof(Guid))
+                    autoValueProperty.PropertyInfo.SetValue(entity, Guid.NewGuid());
+                else if (keyInsert && autoValueType == typeof(int))
+                {
+                    insertQuery.AutoIncrementKey = true;
+                    insertQuery.AutoIncrementField = autoValueProperty as SqlKeyPropertyMetadata;
+                }
+                else
+                    throw new NotImplementedException($"AutoValue for type {autoValueProperty.PropertyInfo.PropertyType} is not supported");
+            }
         }
     }
 }