+++ tests/Models/Spaceship.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations;
 using Extended.Dapper.Core.Attributes.Entities;
 using Extended.Dapper.Core.Attributes.Entities.Relations;
@@ -9,6 +10,9 @@ namespace Extended.Dapper.Tests.Models
         [Key]
         [AutoValue]
         public int Id { get; set; }
+
+        [AutoValue]
+        public Guid ExternalId { get; set; }
 
         [ManyToOne(typeof(Author), "OwnerId", true)]
         public Author Owner { get; set; }